+++ src/main/scala/edu/ie3/simona/agent/grid/PowerFlowParams.scala
@@ -20,8 +20,6 @@ import scala.concurrent.duration.FiniteDuration
   *   equations
   * @param maxIterations
   *   Maximum amount of iterations in a [[edu.ie3.powerflow.NewtonRaphsonPF]]
-  * @param sweepTimeout
-  *   [[org.apache.pekko.pattern.ask]] timeout for a sweep
   * @param stopOnFailure
   *   Whether to stop the GridAgent after a power flow calculation failed
   */
@@ -29,6 +27,5 @@ final case class PowerFlowParams(
     maxSweepPowerDeviation: Double,
     epsilon: Vector[Double],
     maxIterations: Int,
-    sweepTimeout: FiniteDuration,
     stopOnFailure: Boolean,
 )