+++ src/main/scala/edu/ie3/simona/service/weather/WeatherService.scala
@@ -22,11 +22,11 @@ import edu.ie3.simona.service.ServiceStateData.{
   InitializeServiceStateData,
   ServiceBaseStateData,
 }
-import edu.ie3.simona.service.SimonaService
+import edu.ie3.simona.service.{DataTimeType, SimonaService}
 import edu.ie3.simona.service.weather.WeatherSource.WeightedCoordinates
 import edu.ie3.simona.util.TickUtil.RichZonedDateTime
 import edu.ie3.simona.util.{Coordinate, SimonaConstants}
-import edu.ie3.util.scala.collection.immutable.SortedDistinctSeq
+import edu.ie3.util.scala.collection.immutable.ActivationTickQueue
 import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import org.slf4j.Logger
@@ -50,12 +50,9 @@ object WeatherService extends SimonaService {
     *
     * @param coordinate
     *   The coordinate to register weather data for.
-    * @param dataType
-    *   The type of weather data to register for.
     */
   final case class WeatherRegistrationData(
-      coordinate: Coordinate,
-      dataType: WeatherDataType,
+      coordinate: Coordinate
   )
 
   /** Container storing registered actors for a coordinate.
@@ -66,7 +63,7 @@ object WeatherService extends SimonaService {
     *   Weights mapping surrounding coordinates onto the registered coordinate.
     */
   final case class CoordinateData(
-      registeredActors: Map[WeatherDataType, Set[
+      registeredActors: Map[DataTimeType, Set[
         ActorRef[ServiceMessage.Response]
       ]],
       coordinateWeights: WeightedCoordinates,
@@ -78,11 +75,8 @@ object WeatherService extends SimonaService {
     *   The weather source to retrieve information from.
     * @param coordinateData
     *   A map of the requested coords to their receiving actor references.
-    * @param maybeNextActivationTick
-    *   The next tick at which the service wants to be activated.
     * @param activationTicks
-    *   A sorted set of ticks, that yet have been sent to the scheduler (w\o
-    *   next tick).
+    *   A queue of future ticks that the service will be activated at.
     * @param startDateTime
     *   The simulation time at which simulation started.
     * @param amountOfInterpolationCoords
@@ -91,8 +85,7 @@ object WeatherService extends SimonaService {
   final case class WeatherBaseStateData(
       weatherSource: WeatherSource,
       coordinateData: Map[Coordinate, CoordinateData] = Map.empty,
-      maybeNextActivationTick: Option[Long],
-      activationTicks: SortedDistinctSeq[Long] = SortedDistinctSeq.empty,
+      activationTicks: ActivationTickQueue = ActivationTickQueue.empty,
       startDateTime: ZonedDateTime,
       amountOfInterpolationCoords: Int = 4,
   ) extends ServiceBaseStateData
@@ -122,25 +115,24 @@ object WeatherService extends SimonaService {
         val weatherSource = WeatherSource(sourceDefinition)
 
         /* What is the first tick to be triggered for? And what are further activation ticks */
-        val (maybeNextTick, furtherActivationTicks) = SortedDistinctSeq(
+        val activationTicks = ActivationTickQueue(
           weatherSource
             .getDataTicks(
               SimonaConstants.FIRST_TICK_IN_SIMULATION,
               simulationEnd.toTick,
             )
             .toSeq
-        ).pop
+        )
 
         val weatherInitializedStateData = WeatherBaseStateData(
           weatherSource,
-          activationTicks = furtherActivationTicks,
-          maybeNextActivationTick = maybeNextTick,
+          activationTicks = activationTicks,
           startDateTime = startDateTime,
         )
 
         Success(
           weatherInitializedStateData,
-          maybeNextTick,
+          activationTicks.nextTick,
         )
 
       case invalidData =>
@@ -160,13 +152,14 @@ object WeatherService extends SimonaService {
     registrationMessage match {
       case SecondaryServiceRegistrationMessage(
             agentToBeRegistered,
-            WeatherRegistrationData(coordinate, dataType),
+            dataTimeType,
+            WeatherRegistrationData(coordinate),
           ) =>
         Success(
           handleRegistrationRequest(
             agentToBeRegistered,
             coordinate,
-            dataType,
+            dataTimeType,
           )
         )
       case invalidMessage =>
@@ -185,7 +178,7 @@ object WeatherService extends SimonaService {
     *   The agent that wants to be registered.
     * @param coordinate
     *   The coordinate of the agent to be registered.
-    * @param dataType
+    * @param dataTimeType
     *   The weather data type that the agent wants to receive.
     * @param serviceStateData
     *   The current service state data of this service.
@@ -196,7 +189,7 @@ object WeatherService extends SimonaService {
   private def handleRegistrationRequest(
       agentToBeRegistered: ActorRef[ServiceMessage.Response],
       coordinate: Coordinate,
-      dataType: WeatherDataType,
+      dataTimeType: DataTimeType,
   )(using
       serviceStateData: WeatherBaseStateData,
       ctx: ActorContext[Message],
@@ -209,7 +202,7 @@ object WeatherService extends SimonaService {
     )
 
     // collate the provided coordinates into a single entity
-    val registrationResponse = serviceStateData.maybeNextActivationTick
+    val registrationResponse = serviceStateData.activationTicks.nextTick
       .map(RegistrationSuccessfulMessage(ctx.self, _))
       .getOrElse(RegistrationFailedMessage(ctx.self))
 
@@ -224,7 +217,7 @@ object WeatherService extends SimonaService {
             agentToBeRegistered ! registrationResponse
 
             val coordinateData = CoordinateData(
-              registeredActors = Map(dataType -> Set(agentToBeRegistered)),
+              registeredActors = Map(dataTimeType -> Set(agentToBeRegistered)),
               coordinateWeights = weightedCoordinates,
             )
 
@@ -245,7 +238,7 @@ object WeatherService extends SimonaService {
 
       case Some(coordinateData) =>
         val registeredActors =
-          coordinateData.registeredActors.getOrElse(dataType, Set.empty)
+          coordinateData.registeredActors.getOrElse(dataTimeType, Set.empty)
 
         if registeredActors.contains(agentToBeRegistered) then
           ctx.log.warn(
@@ -256,7 +249,7 @@ object WeatherService extends SimonaService {
 
         val adaptedCoordinateData = coordinateData.copy(registeredActors =
           coordinateData.registeredActors +
-            (dataType -> registeredActors.incl(agentToBeRegistered))
+            (dataTimeType -> registeredActors.incl(agentToBeRegistered))
         )
 
         serviceStateData.copy(
@@ -275,13 +268,10 @@ object WeatherService extends SimonaService {
     given simulationStart: ZonedDateTime = serviceStateData.startDateTime
 
     /* Pop the next activation tick and update the state data */
-    val (
-      maybeNextTick: Option[Long],
-      updatedStateData: WeatherBaseStateData,
-    ) = {
-      val (nextTick, remainderTicks) = serviceStateData.activationTicks.pop
-      (nextTick, serviceStateData.copy(activationTicks = remainderTicks))
-    }
+    val remainingTicks = serviceStateData.activationTicks.dropFirst
+    val maybeNextTick = remainingTicks.nextTick
+    val updatedStateData =
+      serviceStateData.copy(activationTicks = remainingTicks)
 
     // get the weather and send it to the subscribed agents
     // no sanity check needed here as we can assume that we always have weather available
@@ -289,11 +279,11 @@ object WeatherService extends SimonaService {
     updatedStateData.coordinateData.foreach { case (_, coordinateData) =>
       val coordinateWeights = coordinateData.coordinateWeights
 
-      coordinateData.registeredActors.foreach { case (dataType, actors) =>
-        val weatherData = dataType match {
-          case WeatherDataType.Current =>
+      coordinateData.registeredActors.foreach { case (dataTimeType, actors) =>
+        val weatherData = dataTimeType match {
+          case DataTimeType.Current =>
             updatedStateData.weatherSource.getWeather(tick, coordinateWeights)
-          case WeatherDataType.CurrentAndForecast(length, interval) =>
+          case DataTimeType.CurrentAndForecast(length, interval) =>
             val endTick = tick + length.toSeconds.toLong
             // weather time series is forwarded without adding error
             val series = updatedStateData.weatherSource