+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentIT.scala
@@ -26,11 +26,11 @@ import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
 import edu.ie3.simona.ontology.messages.ResultMessage.RequestResult
 import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.Data.SecondaryData.WeatherData
-import edu.ie3.simona.service.ServiceType
+import edu.ie3.simona.service.{DataTimeType, ServiceType}
 import edu.ie3.simona.service.primary.PrimaryServiceProxy
 import edu.ie3.simona.service.results.ResultServiceProxy.ExpectResult
 import edu.ie3.simona.service.weather.WeatherService.WeatherRegistrationData
-import edu.ie3.simona.service.weather.{WeatherDataType, WeatherService}
+import edu.ie3.simona.service.weather.WeatherService
 import edu.ie3.simona.test.common.TestSpawnerTyped
 import edu.ie3.simona.test.common.input.EmInputTestData
 import edu.ie3.simona.test.matchers.QuantityMatchers
@@ -121,11 +121,10 @@ class EmAgentIT
         lockActivation ! Activation(PRE_INIT_TICK)
 
         val emAgent = spawn(
-          EmAgent(
+          EmAgentInit(
             emInput,
             modelConfig,
             outputConfigOn,
-            "PRIORITIZED",
             simulationStartDate,
             parent = Left(scheduler.ref),
             listener = resultServiceProxy.ref,
@@ -201,12 +200,12 @@ class EmAgentIT
         weatherService.expectMessage(
           SecondaryServiceRegistrationMessage(
             pvAgent,
+            DataTimeType.Current,
             WeatherRegistrationData(
               Coordinate(
                 pvInput.getNode.getGeoPosition.getY,
                 pvInput.getNode.getGeoPosition.getX,
-              ),
-              WeatherDataType.Current,
+              )
             ),
           )
         )
@@ -409,11 +408,10 @@ class EmAgentIT
         lockActivation ! Activation(PRE_INIT_TICK)
 
         val emAgent = spawn(
-          EmAgent(
+          EmAgentInit(
             emInput,
             modelConfig,
             outputConfigOn,
-            "PRIORITIZED",
             simulationStartDate,
             parent = Left(scheduler.ref),
             listener = resultServiceProxy.ref,
@@ -489,12 +487,12 @@ class EmAgentIT
         weatherService.expectMessage(
           SecondaryServiceRegistrationMessage(
             pvAgent,
+            DataTimeType.Current,
             WeatherRegistrationData(
               Coordinate(
                 pvInput.getNode.getGeoPosition.getY,
                 pvInput.getNode.getGeoPosition.getX,
-              ),
-              WeatherDataType.Current,
+              )
             ),
           )
         )
@@ -508,12 +506,12 @@ class EmAgentIT
         weatherService.expectMessage(
           SecondaryServiceRegistrationMessage(
             hpAgent,
+            DataTimeType.Current,
             WeatherRegistrationData(
               Coordinate(
                 adaptedHpInputModel.getNode.getGeoPosition.getY,
                 adaptedHpInputModel.getNode.getGeoPosition.getX,
-              ),
-              WeatherDataType.Current,
+              )
             ),
           )
         )
@@ -978,11 +976,10 @@ class EmAgentIT
         lockActivation ! Activation(PRE_INIT_TICK)
 
         val emAgent = spawn(
-          EmAgent(
+          EmAgentInit(
             emInput,
             modelConfig,
             outputConfigOn,
-            "PRIORITIZED",
             simulationStartDate,
             parent = Left(scheduler.ref),
             listener = resultServiceProxy.ref,
@@ -1050,12 +1047,12 @@ class EmAgentIT
         weatherService.expectMessage(
           SecondaryServiceRegistrationMessage(
             pvAgent,
+            DataTimeType.Current,
             WeatherRegistrationData(
               Coordinate(
                 pvInputLimitedOperationTime.getNode.getGeoPosition.getY,
                 pvInputLimitedOperationTime.getNode.getGeoPosition.getX,
-              ),
-              WeatherDataType.Current,
+              )
             ),
           )
         )