+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentSpec.scala
@@ -71,11 +71,10 @@ class EmAgentSpec
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
-        EmAgent(
+        EmAgentInit(
           emInput,
           EmRuntimeConfig(),
           outputConfig,
-          "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
           listener = resultProxy.ref,
@@ -260,11 +259,10 @@ class EmAgentSpec
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
-        EmAgent(
+        EmAgentInit(
           emInput,
           EmRuntimeConfig(),
           outputConfig,
-          "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
           listener = resultProxy.ref,
@@ -444,11 +442,10 @@ class EmAgentSpec
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
       val emAgent = spawn(
-        EmAgent(
+        EmAgentInit(
           emInput,
           EmRuntimeConfig(),
           outputConfig,
-          "PRIORITIZED",
           simulationStart,
           parent = Left(scheduler.ref),
           listener = resultProxy.ref,
@@ -639,11 +636,10 @@ class EmAgentSpec
       val parentEmAgent = TestProbe[FlexResponse]("ParentEmAgent")
 
       val emAgent = spawn(
-        EmAgent(
+        EmAgentInit(
           emInput,
           EmRuntimeConfig(),
           outputConfig,
-          "PRIORITIZED",
           simulationStart,
           parent = Right(parentEmAgent.ref),
           listener = resultProxy.ref,