+++ src/test/scala/edu/ie3/simona/agent/grid/DBFSAlgorithmParticipantSpec.scala
@@ -172,6 +172,7 @@ class DBFSAlgorithmParticipantSpec
       // we now answer the request of our gridAgentsWithParticipants
       // with a fake slack voltage message
       firstSlackVoltageRequestSender ! SlackVoltageResponse(
+        superiorGridAgent.ref,
         firstSweepNo,
         Seq(
           ExchangeVoltage(
@@ -196,6 +197,7 @@ class DBFSAlgorithmParticipantSpec
         Seq(
           ExchangePower(
             supNodeA.getUuid,
+            gridAgentWithParticipants,
             Megawatts(135.90837346741768),
             Megavars(60.98643348675892),
           )
@@ -217,6 +219,7 @@ class DBFSAlgorithmParticipantSpec
 
       // the superior grid would answer with updated slack voltage values
       secondSlackAskSender ! SlackVoltageResponse(
+        superiorGridAgent.ref,
         secondSweepNo,
         Seq(
           ExchangeVoltage(
@@ -233,6 +236,7 @@ class DBFSAlgorithmParticipantSpec
         Seq(
           ExchangePower(
             supNodeA.getUuid,
+            gridAgentWithParticipants,
             Megawatts(135.90837346741768),
             Megavars(60.98643348675892),
           )