+++ src/test/scala/edu/ie3/simona/agent/grid/PowerFlowSupportSpec.scala
@@ -84,11 +84,12 @@ class PowerFlowSupportSpec
           )
         )
       ),
-      nodeToReceivedPower = Map(
+      nodeToReceivedGridPower = Map(
         node1.uuid -> Map(
           actorRef -> Some(
             ExchangePower(
               node1.uuid,
+              actorRef,
               Megawatts(1d),
               Megavars(0d),
             )
@@ -348,7 +349,6 @@ class PowerFlowSupportSpec
       1e-5,
       Vector(1e-12),
       50,
-      30.minutes,
       stopOnFailure = true,
     )