+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -166,7 +166,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 6 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2))
             q should approximate(Kilovars(0.968644209676))
         }
@@ -180,7 +180,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         )
 
         gridAgent.expectMessageType[AssetPowerUnchangedMessage] match {
-          case AssetPowerUnchangedMessage(p, q) =>
+          case AssetPowerUnchangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2))
             q should approximate(Kilovars(0.968644209676))
         }
@@ -194,7 +194,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         )
 
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2))
             // not voltage dependent
             q should approximate(Kilovars(0.968644209676))
@@ -242,7 +242,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         )
 
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(4))
             q should approximate(Kilovars(1.93728841935))
         }
@@ -330,7 +330,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 6 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2))
             q should approximate(Kilovars(0.968644209676))
         }
@@ -397,7 +397,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         )
 
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(4))
             q should approximate(Kilovars(1.93728841935))
         }
@@ -544,7 +544,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 6+3=9 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(3))
             q should approximate(Kilovars(1.4529663145))
         }
@@ -679,7 +679,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 6 hours of 6+6=12 kW, 2 hours of 6+9=15 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(8.5))
             q should approximate(Kilovars(4.116737891123))
         }
@@ -815,7 +815,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 3 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(1))
             q should approximate(Kilovars(0.48432210484))
         }
@@ -916,7 +916,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 6 hours of 6 kW, 2 hours of 3 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(3.5))
             q should approximate(Kilovars(1.695127366932))
         }
@@ -1046,7 +1046,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 3 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(1))
             q should approximate(Kilovars(0.48432210483))
         }
@@ -1120,7 +1120,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 3 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2))
             q should approximate(Kilovars(0.96864420966))
         }
@@ -1248,7 +1248,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         )
 
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(1))
             q should approximate(Kilovars(0.48432210483))
         }
@@ -1381,7 +1381,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 1 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(0.6666666667))
             q should approximate(Kilovars(0.32288140322))
         }
@@ -1609,7 +1609,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 3 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(1))
             q should approximate(Kilovars(0.48432210483))
         }
@@ -1827,7 +1827,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 6 hours of 3 kW, 2 hours of 6 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(2.5))
             q should approximate(Kilovars(1.210805262))
         }
@@ -2039,7 +2039,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 8 hours of 0 kW, 4 hours of 3 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(1))
             q should approximate(Kilovars(0.48432210483))
         }
@@ -2247,7 +2247,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         // 6 hours of 6 kW, 2 hours of 3 kW, 4 hours of 0 kW
         gridAgent.expectMessageType[AssetPowerChangedMessage] match {
-          case AssetPowerChangedMessage(p, q) =>
+          case AssetPowerChangedMessage(_, p, q) =>
             p should approximate(Kilowatts(3.5))
             q should approximate(Kilovars(1.695127366932))
         }